+++ lib/runners/processor.rb
@@ -190,26 +190,6 @@ module Runners
       end
     end
 
-    def add_warning_if_deprecated_options
-      name = :options
-      return unless config_linter[name]
-
-      file = config.path_name
-      add_warning <<~MSG, file: file
-        DEPRECATION WARNING!!!
-        The `#{config_field_path(name)}` option is deprecated. Fix your `#{file}` as follows:
-        See #{analyzer_doc} for details.
-
-        ```diff
-         linter:
-           #{analyzer_id}:
-        -    options:
-        -      foo: "bar"
-        +    foo: "bar"
-        ```
-      MSG
-    end
-
     def add_warning_for_deprecated_option(name, to:)
       return unless config_linter[name]