+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Loosen unsatisfied constraints for npm packages [#2171](https://github.com/sider/runners/pull/2171)
 - Improve analysis finish message [#2187](https://github.com/sider/runners/pull/2187)
 - Support npm 7 [#2189](https://github.com/sider/runners/pull/2189)
+- Remove deprecated `linter.{id}.options` [#2190](https://github.com/sider/runners/pull/2190)
 
 ## 0.45.0